+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Please clear changelog after each release.
 Put the changelog BELOW the dashes. ANYTHING ABOVE IS IGNORED.
 -----------------
-- FrozenLib now supports the `jsonc` file extension for configs.
-- FrozenLib's supported `Fabric Language Kotlin` version has been updated to 1.13.7.
-- FrozenLib's embedded `xjs-data` and `xjs-compat` libraries have been updated to versions 0.14 and 0.13 respectively.
+- Added pushable Block Entities!
+  - Blocks within the `frozenlib:has_pushable_block_entity` block tag that have a Block Entity will now be pushable.
+  - The models of Block Entities will render whilst being pushed.
+  - Please note this feature cannot be considered stable, so do not expect to be able to add just any Block to the tag without possible side-effects.